move functions defined in app/src/models/status out to better locations to represent their purpose
