+++ app/src/ui/octicons/index.ts
@@ -1,3 +1,4 @@
 export { Octicon } from './octicon'
 export { OcticonSymbol } from './octicons.generated'
 export { iconForRepository } from './repository'
+export { iconForStatus } from './status'
